requirements.txt
python-magic-bin
dotenv
fastapi
uvicorn
pydantic
python-multipart
requests
opencv-python-headless
ultralytics
numpy<2
boto3
psycopg2-binary
alembic
sqlalchemy
